% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.blockmatrix.matrix.R
\name{as.blockmatrix}
\alias{as.blockmatrix}
\alias{as.blockmatrix.blockmatrix}
\alias{as.blockmatrix.default}
\alias{as.blockmatrix.matrix}
\alias{as.blockmatrix.numeric}
\title{Coerces to a Blockmatrix}
\usage{
as.blockmatrix(M = NULL, ...)

\method{as.blockmatrix}{default}(M, adjust_zero = TRUE, zero_element = "0",
  ...)

\method{as.blockmatrix}{blockmatrix}(M, adjust_zero = TRUE,
  add_zero_matrix = FALSE, zero_element = "0", ...)

\method{as.blockmatrix}{matrix}(M, nrowe = nrow(M), ncole = ncol(M),
  nrow = NULL, ncol = NULL, adjust_zero = TRUE, zero_element = "0", ...)

\method{as.blockmatrix}{default}(M, adjust_zero = TRUE, zero_element = "0",
  ...)

\method{as.blockmatrix}{numeric}(M, adjust_zero = TRUE, zero_element = "0",
  ...)
}
\arguments{
\item{M}{a \code{matrix} or \code{blockmatrix} object}

\item{...}{further arguments}

\item{adjust_zero}{logical value. If \code{TRUE} (Default) it replaces the zero matrices with \code{zero_element}.}

\item{zero_element}{see \code{\link{ncol_elements}} or \code{\link{nrow_elements}}}

\item{add_zero_matrix}{logical value.  If \code{TRUE} it adds a zero-element element matrix as an object called \code{zero_element} in the blockmatrix}

\item{nrowe}{number of rows for each block (element of the blockmatrix)}

\item{ncole}{number of columns for each block (element of the blockmatrix)}

\item{nrow}{number of rows for block-matrix}

\item{ncol}{number of columns of blockmatrix}
}
\description{
Coerces to a Blockmatrix

\code{as.blockmatrix} S3 method for \code{blockmatrix}, \code{matrix} and \code{NULL} object
}
\details{
In case of coercion from a matrix, \code{ncole} and/or \code{ncol} must be a submultiple of the number of columns of \code{M}; \code{nrowe} and/or \code{nrow} must be a submultiple of the number of rows of \code{M}.
}
\author{
Emanuele Cordano
}

