% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.matrix.blockmatrix.R
\name{as.matrix.blockmatrix}
\alias{as.matrix}
\alias{as.matrix.blockmatrix}
\title{\code{as.matrix} S3 method for \code{blockmatrix} object}
\usage{
\method{as.matrix}{blockmatrix}(x, zero_element = "0", ...)
}
\arguments{
\item{x}{a \code{blockmatrix} object}

\item{zero_element}{(see \code{\link{ncol_elements}} or \code{\link{nrow_elements}})}

\item{...}{further arguments (see \code{\link{ncol_elements}} or \code{\link{nrow_elements}})}
}
\description{
\code{as.matrix} S3 method for \code{blockmatrix} object
}
\author{
Emanuele Cordano
}

