% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dim.blockmatrix.R
\name{dim.blockmatrix}
\alias{dim}
\alias{dim.blockmatrix}
\title{\code{dim} S3 method for \code{blockmatrix} object}
\usage{
\method{dim}{blockmatrix}(x)
}
\arguments{
\item{x}{a \code{blockmatrix} object}
}
\description{
\code{dim} S3 method for \code{blockmatrix} object
}
\author{
Emanuele Cordano
}

