% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zbracket[.blockmatrix.R
\name{[.blockmatrix}
\alias{Extract}
\alias{[}
\alias{[.blockmatrix}
\title{\code{[} S3 method for \code{\link{blockmatrix}} object}
\usage{
\method{[}{blockmatrix} (M, i = 1:nrow(M), j = 1:ncol(M),numeric_value=TRUE,blockmatrix=FALSE,...)
}
\arguments{
\item{M}{a \code{blockmatrix} object}

\item{i,j}{matrix indices (numerical or character)}

\item{numeric_value}{logical value . If \code{TRUE} (Default if \code{i,j} have both length 1) and \code{i,j} have both length 1, a \code{i,j} numeric matrix is returened.}

\item{blockmatrix}{logical value. If \code{TRUE} (Default if \code{i} or \code{j} have length greater than 1) a \code{blockmatrix} is returned.}

\item{...}{further argument for \code{\link{[}} method}
}
\value{
The \code{i,j} matrix as a numarical matrix if \code{blockmatrix} is \code{FALSE}, otherwise the returen oblect is a \code{\link{blockmatrix}} object.
In case \code{i} is a character vector, the method returns a list of objects with name containing in \code{i} and taken from \code{M}.
}
\description{
\code{[} S3 method for \code{\link{blockmatrix}} object
}
\author{
Emanuele Cordano
}

