% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/is.zero.blockmatrix.R
\name{is.zero.blockmatrix}
\alias{is.zero.blockmatrix}
\title{is.zero.bolockmatrix}
\usage{
is.zero.blockmatrix(M, not.a.blockmatrix = FALSE)
}
\arguments{
\item{M}{a \code{blockmatrix} object}

\item{not.a.blockmatrix}{value to be returned in case \code{M} is not a a \code{blockmatrix} object}
}
\value{
logical value in case \code{M} is a zero blockmatrix
}
\description{
is.zero.bolockmatrix
}
\author{
Emanuele Cordano
}

