% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/length.blockmatrix.R
\name{length.blockmatrix}
\alias{length}
\alias{length.blockmatrix}
\title{\code{length} S3 method for \code{blockmatrix} object}
\usage{
\method{length}{blockmatrix}(x)
}
\arguments{
\item{x}{a \code{blockmatrix} object}
}
\description{
\code{length} S3 method for \code{blockmatrix} object
}
\author{
Emanuele Cordano
}

