% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.blockmatrix.R
\name{Math.blockmatrix}
\alias{Math}
\alias{Math.blockmatrix}
\alias{Ops}
\alias{Ops.blockmatrix}
\title{\code{Math} and \code{Ops} group of S3 methods for \code{blockmatrix} object}
\usage{
\method{Math}{blockmatrix}(x, as.blockmatrix = TRUE, ...)

\method{Ops}{blockmatrix}(e1, e2)
}
\arguments{
\item{x,e1,e2}{\code{blockmatrix} objects}

\item{as.blockmatrix}{logical value. If \code{TRUE} (Default), the output is a \code{blockmatrix} object}

\item{...}{further arguments}
}
\description{
\code{Math} and \code{Ops} group of S3 methods for \code{blockmatrix} object
}
\author{
Emanuele Cordano
}

