% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/names.replacement.blockmatrix.R
\name{names<-.blockmatrix}
\alias{names<-}
\alias{names<-.blockmatrix}
\title{\code{names<-} S3 Replacement method for \code{blockmatrix} object}
\usage{
\method{names}{blockmatrix} (x) <- value
}
\arguments{
\item{x}{a \code{blockmatrix} object}

\item{value}{character vector with the new names replacing the old names}
}
\value{
\code{x} with new names and replaced names in the \code{value} matrix.
}
\description{
\code{names<-} S3 Replacement method for \code{blockmatrix} object
}
\examples{
library(blockmatrix)

M <- as.blockmatrix(array(rnorm(100),c(10,10)),nrow=2,ncol=2)


pn <- paste("p",names(M),sep="")
names(M) <- pn
}
\author{
Emanuele Cordano
}

