% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ncol_elements.blockmatrix.R
\name{ncol_elements}
\alias{ncol_elements}
\alias{ncol_elements.blockmatrix}
\alias{ncol_elements.default}
\title{\code{ncol_elements} S3 method for \code{blockmatrix} object}
\usage{
ncol_elements(M, zero_element = "0", ...)

\method{ncol_elements}{default}(M, zero_element = "0", ...)

\method{ncol_elements}{blockmatrix}(M, zero_element = "0", ...)
}
\arguments{
\item{M}{a \code{blockmatrix} object}

\item{zero_element}{character value indicating a zero matrix. Default is \code{"0"}}

\item{...}{further arguments}
}
\value{
The number of columns of a matrix-type element of \code{M}. It is \code{NA} if the elements has different number of columns.
}
\description{
\code{ncol_elements} S3 method for \code{blockmatrix} object



\code{ncol_elements} S3 method for \code{blockmatrix} object
}
\author{
Emanuele Cordano
}

