% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nrow.blockmatrix.R
\name{nrow.blockmatrix}
\alias{nrow}
\alias{nrow.blockmatrix}
\title{\code{nrow} S3 method for \code{blockmatrix} object}
\usage{
\method{nrow}{blockmatrix}(M)
}
\arguments{
\item{M}{a \code{blockmatrix} object}
}
\value{
Number of rows of blockmatrix \code{M}
}
\description{
\code{nrow} S3 method for \code{blockmatrix} object
}
\author{
Emanuele Cordano
}

