% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nrow_elements.blockmatrix.R
\name{nrow_elements}
\alias{nrow_elements}
\alias{nrow_elements.blockmatrix}
\alias{nrow_elements.default}
\title{\code{nrow_elements} S3 method for \code{blockmatrix} object}
\usage{
nrow_elements(M, zero_element = "0", ...)

\method{nrow_elements}{default}(M, zero_element = "0", ...)

\method{nrow_elements}{blockmatrix}(M, zero_element = "0", ...)
}
\arguments{
\item{M}{a \code{blockmatrix} object}

\item{zero_element}{character value indicating a zero matrix. Default is \code{"0"}}

\item{...}{further arguments}
}
\value{
The number of rows of a matrix-type element of \code{M}. It is \code{NA} if the elements has different number of rows.
}
\description{
\code{nrow_elements} S3 method for \code{blockmatrix} object



\code{nrow_elements} S3 method for \code{blockmatrix} object
}
\author{
Emanuele Cordano
}

