% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/t.blockmatrix.R
\name{t.blockmatrix}
\alias{t}
\alias{t.blockmatrix}
\title{\code{t} 'transpose' S3 method for \code{blockmatrix} object}
\usage{
\method{t}{blockmatrix}(x)
}
\arguments{
\item{x}{a \code{blockmatrix} object}
}
\description{
\code{t} 'transpose' S3 method for \code{blockmatrix} object
}
\author{
Emanuele Cordano
}

