\name{blockmatmult}
\alias{blockmatmult}
\title{\code{blockmatmult} implements the implents betwwen two blockmatrix ( see \code{\link{matmult}} for \code{matrx} objects)}
\usage{
  blockmatmult(x, y, ...)
}
\arguments{
  \item{x,y}{\code{blockmatrix} objects}

  \item{...}{further arguments}
}
\value{
  The inner product between \code{x} and \code{y} as a
  \code{blockmatrix} object
}
\description{
  \code{blockmatmult} implements the implents betwwen two
  blockmatrix ( see \code{\link{matmult}} for \code{matrx}
  objects)
}
\author{
  Emanuele Cordano
}

