\name{ncol.blockmatrix}
\alias{ncol}
\alias{ncol.blockmatrix}
\title{\code{ncol} S3 method for \code{blockmatrix} object}
\usage{
  \method{ncol}{blockmatrix} (M)
}
\arguments{
  \item{M}{a \code{blockmatrix} object}
}
\value{
  Numbner of columns of blockmatrix \code{M}
}
\description{
  \code{ncol} S3 method for \code{blockmatrix} object
}
\author{
  Emanuele Cordano
}

