\name{t.blockmatrix}
\alias{t}
\alias{t.blockmatrix}
\title{\code{t} 'transpose' S3 method for \code{blockmatrix} object}
\usage{
  \method{t}{blockmatrix} (x)
}
\arguments{
  \item{x}{a \code{blockmatrix} object}
}
\description{
  \code{t} 'transpose' S3 method for \code{blockmatrix}
  object
}
\author{
  Emanuele Cordano
}

